linter(no-cylce): moved createChatSignature from helper.ts to create.ts
