+++ common/chat/create.ts
@@ -1,12 +1,14 @@
 import { User } from '../user';
 import { ChatMetaData, ContactReason, Conversation, ConversationInfos, FinishedReason, SystemMessage, TJConversation } from './types';
-import { checkChatMembersAccessRights, convertTJConversation, createOneOnOneId } from './helper';
+import { checkChatMembersAccessRights, convertTJConversation, createOneOnOneId, userIdToTalkJsId } from './helper';
 import { createConversation, getConversation, markConversationAsWriteable, sendSystemMessage, updateConversation } from './conversation';
 import { getOrCreateChatUser } from './user';
 import { prisma } from '../prisma';
 import { getMyContacts } from './contacts';
 import systemMessages from './localization';
 import { getLogger } from '../logger/logger';
+import assert from 'assert';
+import { createHmac } from 'crypto';
 
 const logger = getLogger('Chat');
 const getOrCreateOneOnOneConversation = async (
@@ -62,6 +64,15 @@ async function ensureChatUsersExist(participants: [User, User] | User[]): Promis
         })
     );
 }
+
+const createChatSignature = async (user: User): Promise<string> => {
+    const TALKJS_SECRET_KEY = process.env.TALKJS_API_KEY;
+    assert(TALKJS_SECRET_KEY, `No TalkJS secret key to create a chat signature for user ${user.userID}.`);
+    const userId = (await getOrCreateChatUser(user)).id;
+    const key = TALKJS_SECRET_KEY;
+    const hash = createHmac('sha256', key).update(userIdToTalkJsId(userId));
+    return hash.digest('hex');
+};
 
 async function handleExistingConversation(
     conversationId: string,
@@ -195,4 +206,4 @@ async function createContactChat(meUser: User, contactUser: User): Promise<strin
     return conversation.id;
 }
 
-export { getOrCreateOneOnOneConversation, getOrCreateGroupConversation, createContactChat };
+export { getOrCreateOneOnOneConversation, getOrCreateGroupConversation, createContactChat, createChatSignature };
